+++ run.py
@@ -1,0 +1,4 @@
+from webstudio import *
+
+if __name__ == '__main__':
+    socketio.run(app, debug=True)
