+++ webstudio/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask, render_template, session, request
+from flask_socketio import SocketIO, emit, join_room, leave_room, \
+    close_room, rooms, disconnect
+import os, random, string, pexpect
+import subprocess as sp
+from threading import Thread
+from queue import Queue,Empty
+import time
+async_mode = None
+app = Flask(__name__)
+app.config['SECRET_KEY'] = 'secret!'
+socketio = SocketIO(app, async_mode=async_mode)
+thread = None
+from webstudio import socket
