+++ webstudio/temp/7Y2fEfJLffCITeC.por
@@ -1,0 +1,12 @@
+programa
+{
+	funcao inicio ()
+	{
+		cadeia nome
+
+		escreva("Olá mundo! Em portugol ")
+    escreva("Qual seu nome?")
+		leia(nome)
+		escreva("Seu nome é: ", nome, "\n")
+	}
+}
