+++ webstudio/temp/PKFYi1ikfivBWIC.por
@@ -1,0 +1,11 @@
+programa
+{
+	funcao inicio ()
+	{
+		cadeia nome
+
+		escreva("Olá mundo! Em portugol ")
+    escreva("Qual seu nome?")
+
+	}
+}
